move src/tabular_trees/sklearn/trees.py -> src/tabular_trees/sklearn.py
