--- src/tabular_trees/sklearn/__init__.py
@@ -1,3 +1,0 @@
-"""Module containing functionality for dealing with sklearn models."""
-
-from . import trees
