+++ tests/sklearn/test_export_tree_data__hist_gradient_boosting_model.py
@@ -3,13 +3,13 @@ import re
 import pandas as pd
 import pytest
 
-from tabular_trees.sklearn import trees
+from tabular_trees import sklearn
 
 
 def test_successful_call(sklearn_diabetes_hist_gbm_regressor):
     """Test a successful call to export_tree_data__hist_gradient_boosting_model."""
 
-    trees.export_tree_data__hist_gradient_boosting_model(
+    sklearn.export_tree_data__hist_gradient_boosting_model(
         sklearn_diabetes_hist_gbm_regressor
     )
 
@@ -26,7 +26,7 @@ def test_model_not_correct_type_exception():
 
     with pytest.raises(TypeError, match=re.escape(msg)):
 
-        trees.export_tree_data__hist_gradient_boosting_model(["a", "b"])
+        sklearn.export_tree_data__hist_gradient_boosting_model(["a", "b"])
 
 
 def test_multiple_responses_exception(sklearn_iris_hist_gbm_classifier):
@@ -37,7 +37,7 @@ def test_multiple_responses_exception(sklearn_iris_hist_gbm_classifier):
         NotImplementedError, match="model with multiple responses not supported"
     ):
 
-        trees.export_tree_data__hist_gradient_boosting_model(
+        sklearn.export_tree_data__hist_gradient_boosting_model(
             sklearn_iris_hist_gbm_classifier
         )
 
@@ -68,15 +68,15 @@ def test_output(mocker, sklearn_diabetes_hist_gbm_regressor):
     )
 
     mocker.patch.object(
-        trees, "_extract_hist_gbm_tree_data", return_value=dummy_tree_data
+        sklearn, "_extract_hist_gbm_tree_data", return_value=dummy_tree_data
     )
 
-    exported_trees = trees.export_tree_data__hist_gradient_boosting_model(
+    exported_trees = sklearn.export_tree_data__hist_gradient_boosting_model(
         sklearn_diabetes_hist_gbm_regressor
     )
 
     assert (
-        type(exported_trees) is trees.ScikitLearnHistTabularTrees
+        type(exported_trees) is sklearn.ScikitLearnHistTabularTrees
     ), "output from export_tree_data__hist_gradient_boosting_model incorrect type"
 
     pd.testing.assert_frame_equal(exported_trees.trees, dummy_tree_data)
